requirements.txt
asgiref==3.8.1
Django==5.0.6
djangorestframework==3.15.1
numpy==1.26.4
packaging==24.0
pandas==2.2.2
plotly==5.22.0
python-dateutil==2.9.0.post0
pytz==2024.1
six==1.16.0
sqlparse==0.5.0
tenacity==8.3.0
typing_extensions==4.11.0
tzdata==2024.1
